Move lookahead queue related things to the util package
